anv/xe: adjust to the new return values from the vm_bind ioctl
The xe.ko team announced a plan to change the return values from the
vm_bind ioctl to match what's described in this commit.

What today is just ENOMEM will become ENOMEM or ENOSPC depending on
the reason. For us we'll still handle both cases in the same way: try
to do things slowly and if they still fail, leave it to the upper
layer.

They also announced ENOSPC for the "out of device memory" case, which
we're now appropriately signaling here.

There's still a lot that can be in the "else" case, but those are
most probably bugs in our driver, so just report the device as lost in
the hope that if/when the application retries, the issue will be gone.

Cc: Thomas Hellström <thomas.hellstrom@linux.intel.com>
Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>