vulkan: change STACK_ARRAY to use alloca()
This should save us significant stack space.

Without this patch, no matter what size is passed to STACK_ARRAY,
we'll always define _stack_##name[8]. If size is less than 8, then
we'll waste some space with unused elements. If size is more than 8,
then we'll waste all 8 elements with a completely unused array,
because we'll declare the array anyway but use malloc().

In practice I see the vast vast majority of STACK_ARRAY() used for
games in Anv is of size 1, so most of the times we're wasting 7
elements of space. The main culprit I see for Anv is
vk_drm_syncobj_wait_many(), which always defines 2 STACK_ARRAYs,
leading to 84 extra bytes per call.

So this commit replaces the fixed-size array - that may or may not be
used - with an array (from either alloca() or malloc()) that always
has the correct size. No wasted space, with the only overhead being a
new boolean that's used to check if we need to free() the array.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>