vulkan: change STACK_ARRAY's threshold to 1024 bytes instead of 8 elements
The threshold is what helps us decide between using the stack or the
heap.  It doesn't really make a lot of sense for us to make our
decision of stack vs heap based on the number of elements because that
doesn't take into consideration sizeof(type). What we really want here
is to avoid a stack overflow, so use the actual size of the array in
bytes to make the decision. That means, of course, the number of
elements will be different based on sizeof(type) and not fixed to 8.

The Anv TR-TT code path uses a bunch of STACK_ARRAYs with really small
sizeof(type), so this commit allows us to choose the stack more often.

On Linux the stack is generally 8MB in size, so 1k sounds like a
reasonable threshold, and it's also used by other projects that have
the same kind of data structure as STACK_ARRAY.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>