anv: use A64 pull push constant mechanism for CS/RT/Mesh/Task stages
Previously the loading mechanism was done in NIR but then once
translated into the backend IR, all kind of find_live_channel
operations would be added because we have to figure out the lanes
alive to build A64 addresses.

With the custom backend solution in assign_curb_setup(), we can do the
operations with NoMask and avoid all the live lanes stuff altogether.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>