intel-rt: allow to build it on non-x86 platforms
With the current state of Xe driver, intel-rt actually compiles on aarch64
and can even run for a short while (on my test system, after few minutes
in Q2RTX, kernel driver locked up). It seems to me, there is no real need
to forbid compiling intel-rt on other 64-bit architectures, but as
discussed in the MR comments, I'll add just aarch64 as the system that I
had.