wsi/wayland: Use `wl_fixes` to destroy `wl_registry`
Destroy `wl_registry` objects using
https://gitlab.freedesktop.org/wayland/wayland/-/merge_requests/388.

Not sure how best to make use of unreleased `libwayland` features here.
But anyway, this works as a placeholder, and can be tested by setting
`VK_ICD_FILENAMES`, `LD_LIBRARY_PATH`, and
`__EGL_VENDOR_LIBRARY_FILENAMES`.

I guess it will need to be made conditional libwayland version once
the changes are merged in libwayland. Or the libwayland version with
this feature would be a hard dependency.