brw: rework constant lowering
We have up to 4 chunks of push constants delivered in the EU registers.

Currently we're referring those using a UNIFORM register type with :
   - nr     : identifier to the chunk of push constant
   - offset : offset in the chunk loaded data

After this change, we have all the constant data identified upfront
and layed out a single constant data range in the backend before
applying the NIR->BRW translation. When the backend translate uniform
accesses, it just calls uniform_reg() with the block/offset/length
parameters and that function returns a single UNIFORM register with :
    - offset : offset in the single constant data

In assign_curb_setup() we just have add the UNIFORM offset to the
first register where the constants are located.

The building of a single constant data range upfront helps a lot for
allowing discontinuity in the pull ranges from Gfx12.5+
compute/mesh/task/rt shaders. That way all the constants can be loaded
way in a tight spot rather than leaving some unused GRFs when a 32B
range is unused somewhere in the push constants.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>