nir: Add a pass to split temporary variables
This normally isn't necessary, because most temporaries are lowered by
nir_lower_vars_to_ssa, but if only part of a variable could be lowered
then the entire variable would remain, resulting in variables being
lowered to scratch unnecessarily. In particular this was happening with
ray queries, which in the turnip implementation are lowered to a struct
including an indirectly-accessed stack and a lot of directly-accessed
members describing the state. This pass splits up the ray query
variable so that most of it can removed and the stack can be lowered to
fast indirect variable accesses.