brw: pass msaa_flags param by sources for fb_write/interp
We want to introduce a new opcode for dynamic msaa flags which will be
lower immediately after the NIR->BRW translation. Then we should not
ever use fs_visitor::dynamic_msaa_flags() again.

The reason for this new opcode is that we need to reference the
dynamic msaa flags value prior to constant value analysis.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>