panvk: Allow compressed formats
Compressed formats were disabled because of the differnt layout for
u_interleaved(compressed), where the 16x16 texel tile model for
non-compressed is replaced by a 4x4 block tile. This prevents us from
creating RGBA views of compressed images, which break copies.

The simple option is to keep compressed images linear, the more complex
one would be to patch coordinates/frag-coord shader side when we're
dealing with a compressed texture/color-attachment. Let's pick the first
option for now, and we'll see if we want to revisit it later on.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>