nir: Ensure all variations of 'a == a' are exact
Comparisons 'a == a' and 'a != a' are special. Applications may use
this, without marking them precise, as open-coded version of
isnan(). Ensure that every instance of one of these comparisons is
marked precise. This guarantees that transformations that convert these
to something else will also generate exact comparisons.

No changes in shader-db or fossil-db on any Intel platform.

Suggested-by: Georg Lehmann