nir: Cleanups after previous commit
These things no longer need to be explictily marked exact. I have left
this is a separate commit just in case it causes a problem. It would be
easier to rever this commit than it would be to revert the previous
commit... and fix everything that depends on it.

No changes in shader-db or fossil-db on any Intel platform.