mesa/main: validate integer-formats
RG integer-textures are only supported on OpenGL if the combination
of EXT_texture_integer and ARB_texture_rg is supported. It's also
supported on GL3, but both of those extensions are required there
anyway. In addition GLES3 is supported.

BGR, BGRA and alpha integer-textures are only supported by
EXT_texture_integer.

Luminance and luminance-alpha integer-textures similarly, but are
unsupported in core contexts, because general luminance support is
removed.

Reviewed-by: Adam Jackson <ajax@redhat.com>