iris: Check resolves/flushes when a new program reads different bindings
The pre-draw resolve handling considers the currently bound framebuffer
against textures which are read by the current shader.  Often when
switching shaders, applications will bind new textures or framebuffers,
but they don't technically have to.  They can leave a set of textures
bound and switch between programs that read different subsets of those
shaders.  In this case, we need to recheck for feedback loops.

Francisco Jerez tracked down some rendering corruption on Meteorlake
where two back to back draw calls used the same resources.  The first
had the same image bound as a texture and also as a framebuffer.  The
second kept the same resources bound, but the new shader no longer
read the texture, so there was no longer a feedback loop.

This patch should make iris_predraw_resolve_inputs happen when a new
program changes the subset of textures that it reads.  Together with
the previous patch, this should cause feedback loop detection to kick
in and trigger reconsideration of the render aux modes, and possibly
a render target flush.

Huge thanks to Francisco for tracking this down.