iris: Flag when rendering feedback loops change
There are a variety of scenarios where either the set of textures for a
stage are rebound, or a new shader is bound that reads a different
subset of the currently bound textures.  This can cause a different set
of textures to be compared against the framebuffer, which could cause a
change in whether or not we currently have a rendering feedback loop.

If we detect a new feedback loop, or lack there of, then we need to
reconsider the render aux mode in iris_predraw_resolve_framebuffer().