iris: Detect rendering feedback loops even without aux enabled
Auxiliary buffers are the main issue with rendering feedback loops,
as block compression causes the "non-overlapping regions" test for
disjoint reading/writing to break.  However, if we have a feedback
loop in one draw (both rendering and texturing from the same surface),
then the next draw renders to it but doesn't sample, we still need to
flush to ensure the second draw's writes happen after the first draw's
reads are fully completed.

Thanks to Francisco Jerez for pointing out this issue.