nine: mark the vertex stream when handling DrawIndexedPrimitiveUP
The nine frontend missed setting the vertex stream within the
context->vtxbuf_mask when implementing DrawIndexPrimitiveUP method.
DX9 remarks on this method say that it "supports only a single
vertex stream, which must be declared as stream #0".

The missing state:
(1) causes the update_vertex_elements routine to access
    uninitialized elements of vtxbuf_holes_map, causing further
    such accesses to related/dependent data-structures, and
(2) seems to be a probable cause of some crashes (due to accesses
    to uninitialized values) reported at #10557.

The changes temporarily borrow the context's state in order to
carry out the rendering, while preserving any active, instance-state
set by prior DX9 calls.

The changes also update the status of the r300 tests, as suggested by
Pavel Ondračka.

Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/10557
Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/11317
Fixes: 9168a376924 ("st/nine: Refactor DrawPrimitiveUp")

Tested-by: Pavel Ondračka <pavel.ondracka@gmail.com>