crocus: uncomment the function crocus_clear_depth_stencil
This function is utilized to carry out partial Depth/Stencial clearing. Since
it's functionality is entirely commented out, the apps that depend on it see
improper/null rendering.

Uncommenting the functionality, ensuring that its general structure matches
that of the corresponding function in `iris`, helps solve the
improper/null rendering issue.

Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/11302
Fixes: f3630548f1d ("crocus: initial gallium driver for Intel gfx 4-7")