glsl: Count the number of signatures passed as varargs to add_function.
In C++, NULL might be defined to 0, and using it in a vararg function
might end up with a 32bit 0 value pushed on the stack. Depending on the
stack layout and previous contents, it's then possible that the value
that is read from va_arg as a 64bit pointer isn't 0.