zink: rework zink_get_name() to cleanup vk_DriverId_to_str() calls and handling
Avoid duplicate calls to the enum-to-str function, avoid moving the
pointer past valid memory addresses, and fixes the *prefix* to check to
actually be that (although realistically there will never be a string
that matches without this being the prefix).

I also think the code is easier to read like this 😇

No functional change for real use-cases though.