intel/brw: Allow multiple uses of LOAD_SUBGROUP_INVOCATION
The NIR to BRW conversion created a single LOAD_SUBGROUP_INVOCATION at
the top of the shader and then lowered nir_intrinsic_load_subgroup_invocation
to its result -- stored as part of NIR to BRW state.

As part of moving some of the lowering of other NIR intrinsics later, it
will be necessary to use the subgroup invocation after NIR to BRW.

So remove the policy of single usage of LOAD_SUBGROUP_INVOCATION and let
the lowering write the cached computation at the top of the shader if
needed.  This will make easier in the future to change this approach,
e.g. to use CSE directly.