nvk: Drop codegen support
Now that NAK supports Maxwell+, there's not much point in having a
codegen back-end anymore.  In theory, it might help with Kepler bring-up
some but we really need to add a Kepler back-end for NAK if we want NVK
to work on Kepler.  This lets just just assume NAK going forward and
never have to think about other back-ends again.