glx: Don't try to dlopen ourselves
This can't do anything useful. The gallium driver does not depend on any
symbols exported by the GLX library, so this cannot make the gallium
driver work better. We can only get here by attempting to initialize
GLX on a display, which means the GLX symbols must already be visible
enough to their caller to work.

If we do need to do this anywhere - and I'm really unconvinced we do,
especially for glvnd - we ought to do it earlier, in __glXInitialize.