frontends/va: Destroy surface buffers that are used for DPB in encode
These surfaces are meant to be used directly by encoder as DPB buffers,
but we don't support this because no driver can implement it.
Instead we only use the surface ids in frontend and the buffers are not
accessible from drivers.
Destroy the buffers to save memory. The frontend code checks for NULL
buffers everywhere, so this should not cause any issues.