gallium/hud: Add assert for potential overrun of fixed length string
Coverity has noticed that gr->name is 128 bytes, and that the name
parameter is unbounded, and potentially could overwrite that. In
practice that cannot actually happen, because name is validated ahead of
time. However, adding the assert could catch a potential future issue,
and communicates that we know we're not going to write off the edge of
our memory.