panfrost: implement load/store for image2DMS on valhall
The way images are stored on valhall is slightly different from how
they were done on bifrost and midgard, and that makes the simple
strategy of pretending they are 3D images not work. This patch leaves
the dimension as SAMPLER_DIM_MS and handles that case when emitting
coordinates.

It also contains a tweak to keep array indices (if present) so that
image2DMSArray works on valhall.