bi: Do not mark tex ops as skip when a dest is used
Previously, it was possible to have a texture operation marked as SKIP
while one of the dests was in use. If an helper thread was to execute
that instruction, it would result in an undefined value being used.

This fix
"dEQP-VK.graphicsfuzz.cov-nested-loops-sample-opposite-corners" where
helper threads would get stuck inside a loop depending on the result of
a TEXS_2D invocation.

Signed-off-by: Mary Guillemard <mary.guillemard@collabora.com>